+++ frontend/src/components/chat/code-engine/componentInspector.tsx
@@ -8,7 +8,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
 import { getElementStyles, updateElementStyle, updateElementContent } from './iframe-click-handler';
-import { StyleUpdateService } from './styleUpdateService';
+import { StyleUpdateService } from './style-update';
 import { toast } from 'sonner';
 
 // Type for component data from custom inspector
